+++ src/test/run-pass/u32-decr.rs
@@ -1,8 +1,8 @@
 // -*- rust -*-
 
 fn main() {
-    let u32 word = (200000 as u32);
-    word = word - (1 as u32);
-    check(word == (199999 as u32));
+    let u32 word = (200000u32);
+    word = word - (1u32);
+    check(word == (199999u32));
 }
