+++ src/test/run-pass/u8-incr-decr.rs
@@ -4,9 +4,9 @@
 // in the rest of the generated code so they're easily grep-able.
 
 fn main() {
-  let u8 x = 19 as u8;  // 0x13
-  let u8 y = 35 as u8;  // 0x23
-  x = x + (7 as u8);    // 0x7
-  y = y - (9 as u8);    // 0x9
+  let u8 x = 19u8;  // 0x13
+  let u8 y = 35u8;  // 0x23
+  x = x + (7u8);    // 0x7
+  y = y - (9u8);    // 0x9
   check(x == y);
 }
