+++ src/test/run-pass/u8-incr.rs
@@ -1,12 +1,12 @@
 // -*- rust -*-
 
 fn main() {
-  let u8 x = 12 as u8;
-  let u8 y = 12 as u8;
-  x = x + (1 as u8);
-  x = x - (1 as u8);
+  let u8 x = 12u8;
+  let u8 y = 12u8;
+  x = x + (1u8);
+  x = x - (1u8);
   check(x == y);
-  //x = 14 as u8;
-  //x = x + 1 as u8;
+  // x = 14u8;
+  // x = x + 1u8;
 }
